requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
bcrypt==4.1.1
pydantic==2.5.0
sqlalchemy==2.0.47
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0